refactor: split internal/file/manager.go by domain
Pure move: manager.go (811 lines, 25 declarations) split into
concern-named sibling files (manager_store, manager_query,
manager_records, manager_mutate, manager_paths), following the
same-package sibling-file convention used by internal/config/config_*.go,
internal/message/manager_*.go, and internal/user/manager_*.go. Only
import blocks changed; no behavior, exported API, or test changes.
