+++ internal/file/manager.go
@@ -9,10 +9,6 @@ import (
 	"sort"
 	"strings"
 	"sync"
-	"time"
-
-	"github.com/ViSiON-3/vision-3-bbs/internal/archiver"
-	"github.com/google/uuid"
 )
 
 // FileManager manages file areas and their associated file records.
@@ -123,86 +119,6 @@ func (fm *FileManager) loadAreas() error {
 	return nil
 }
 
-// loadAllFileRecords iterates through loaded areas and loads their metadata.
-func (fm *FileManager) loadAllFileRecords() error {
-	fm.muAreas.RLock() // Need read lock on areas to iterate
-	defer fm.muAreas.RUnlock()
-	fm.muFiles.Lock() // Need write lock on fileRecords map
-	defer fm.muFiles.Unlock()
-
-	fm.fileRecords = make(map[int][]FileRecord) // Reset records
-	var totalFilesLoaded int
-	var errorsEncountered bool
-
-	for areaID, area := range fm.fileAreas {
-		metadataPath := filepath.Join(fm.basePath, area.Path, "metadata.json")
-		data, err := os.ReadFile(metadataPath)
-		if err != nil {
-			if os.IsNotExist(err) {
-				slog.Debug("metadata file not found for area, assuming no files", "path", metadataPath, "area", area.Tag, "id", areaID)
-				fm.fileRecords[areaID] = []FileRecord{} // Initialize empty slice
-				continue
-			}
-			slog.Error("failed to read metadata file for area", "path", metadataPath, "area", area.Tag, "error", err)
-			errorsEncountered = true
-			continue // Skip this area on error
-		}
-
-		var records []FileRecord
-		if err := json.Unmarshal(data, &records); err != nil {
-			slog.Error("failed to parse metadata file for area", "path", metadataPath, "area", area.Tag, "error", err)
-			errorsEncountered = true
-			continue // Skip this area on error
-		}
-
-		// TODO: Validate records? Ensure filenames exist?
-		fm.fileRecords[areaID] = records
-		totalFilesLoaded += len(records)
-		slog.Debug("loaded file records for area", "count", len(records), "area", area.Tag, "id", areaID)
-	}
-
-	slog.Info("loaded metadata for areas", "areas", len(fm.fileAreas), "count", totalFilesLoaded)
-	if errorsEncountered {
-		return fmt.Errorf("encountered errors while loading file records")
-	}
-	return nil
-}
-
-// saveFileRecords saves the metadata for a specific file area.
-func (fm *FileManager) saveFileRecords(areaID int) error {
-	fm.muAreas.RLock() // Need read lock for area path
-	area, exists := fm.fileAreas[areaID]
-	if !exists {
-		fm.muAreas.RUnlock()
-		return fmt.Errorf("cannot save records for non-existent area ID %d", areaID)
-	}
-	metadataPath := filepath.Join(fm.basePath, area.Path, "metadata.json")
-	fm.muAreas.RUnlock()
-
-	fm.muFiles.Lock() // Need write lock to marshal and write
-	defer fm.muFiles.Unlock()
-
-	records, exists := fm.fileRecords[areaID]
-	if !exists {
-		// This might happen if an area was loaded but metadata failed initially
-		// Or if called for a newly created area before any records added.
-		slog.Debug("no records found in memory for area during save, saving empty list", "id", areaID)
-		records = []FileRecord{} // Ensure we save an empty list, not nil
-	}
-
-	data, err := json.MarshalIndent(records, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal file records for area %d: %w", areaID, err)
-	}
-
-	if err := os.WriteFile(metadataPath, data, 0644); err != nil {
-		return fmt.Errorf("failed to write metadata file %s: %w", metadataPath, err)
-	}
-
-	slog.Debug("saved file records for area", "count", len(records), "area", area.Tag, "id", areaID, "path", metadataPath)
-	return nil
-}
-
 // --- Public API Methods --- //
 
 // ListAreas returns a sorted list of FileArea definitions.
@@ -244,568 +160,4 @@ func (fm *FileManager) GetAreaByID(id int) (*FileArea, bool) {
 
 	area, exists := fm.fileAreas[id]
 	return area, exists // Return pointer directly
-}
-
-// GetFileRecordByID looks up a file record by UUID across all areas.
-func (fm *FileManager) GetFileRecordByID(fileID uuid.UUID) (*FileRecord, error) {
-	fm.muFiles.RLock()
-	defer fm.muFiles.RUnlock()
-
-	for _, records := range fm.fileRecords {
-		for i := range records {
-			if records[i].ID == fileID {
-				rec := records[i]
-				return &rec, nil
-			}
-		}
-	}
-	return nil, fmt.Errorf("file record with ID %s not found", fileID)
-}
-
-// SearchFiles returns file records whose filename or description contains
-// query (case-insensitive) across all areas.
-func (fm *FileManager) SearchFiles(query string) []FileRecord {
-	fm.muFiles.RLock()
-	defer fm.muFiles.RUnlock()
-
-	lowerQuery := strings.ToLower(query)
-	var results []FileRecord
-	for _, records := range fm.fileRecords {
-		for _, rec := range records {
-			if strings.Contains(strings.ToLower(rec.Filename), lowerQuery) ||
-				strings.Contains(strings.ToLower(rec.Description), lowerQuery) {
-				results = append(results, rec)
-			}
-		}
-	}
-	return results
-}
-
-// GetFilesNewerThan returns file records in the given area uploaded after since.
-func (fm *FileManager) GetFilesNewerThan(areaID int, since time.Time) []FileRecord {
-	fm.muFiles.RLock()
-	defer fm.muFiles.RUnlock()
-
-	var results []FileRecord
-	for _, rec := range fm.fileRecords[areaID] {
-		if rec.UploadedAt.After(since) {
-			results = append(results, rec)
-		}
-	}
-	return results
-}
-
-// GetUnreviewedFiles returns file records in the given area where Reviewed is false.
-func (fm *FileManager) GetUnreviewedFiles(areaID int) []FileRecord {
-	fm.muFiles.RLock()
-	defer fm.muFiles.RUnlock()
-
-	var results []FileRecord
-	for _, rec := range fm.fileRecords[areaID] {
-		if !rec.Reviewed {
-			results = append(results, rec)
-		}
-	}
-	return results
-}
-
-// GetFilesForArea returns a slice of FileRecord for a given area ID.
-// Returns an empty slice if the area doesn't exist or has no files.
-func (fm *FileManager) GetFilesForArea(areaID int) []FileRecord {
-	fm.muFiles.RLock()
-	defer fm.muFiles.RUnlock()
-
-	records, exists := fm.fileRecords[areaID]
-	if !exists {
-		return []FileRecord{} // Return empty slice, not nil
-	}
-
-	// Return a copy to prevent external modification of the internal slice
-	// TODO: Consider sorting options here (filename, date, etc.)
-	recordsCopy := make([]FileRecord, len(records))
-	copy(recordsCopy, records)
-	return recordsCopy
-}
-
-// GetFileCountForArea returns the total number of file records for a given area ID.
-// Returns 0 if the area doesn't exist or has no files.
-func (fm *FileManager) GetFileCountForArea(areaID int) (int, error) {
-	fm.muFiles.RLock()         // Acquire read lock for accessing file records
-	defer fm.muFiles.RUnlock() // Ensure lock is released
-
-	records, exists := fm.fileRecords[areaID]
-	if !exists {
-		// Area might not exist or simply hasn't had metadata loaded/saved yet.
-		// Returning 0 is appropriate for the caller (runListFiles).
-		slog.Debug("area not found or unloaded; returning file count 0", "id", areaID)
-		return 0, nil
-	}
-
-	return len(records), nil
-}
-
-// GetTotalFileCount returns the total number of files across all areas.
-func (fm *FileManager) GetTotalFileCount() int {
-	areas := fm.ListAreas()
-	total := 0
-	for _, area := range areas {
-		count, err := fm.GetFileCountForArea(area.ID)
-		if err != nil {
-			continue
-		}
-		total += count
-	}
-	return total
-}
-
-// GetFilesForAreaPaginated returns a slice of FileRecord for a given area ID,
-// limited to the specified page and pageSize.
-// Returns an empty slice if the area doesn't exist, has no files, or the page is out of bounds.
-func (fm *FileManager) GetFilesForAreaPaginated(areaID int, page int, pageSize int) ([]FileRecord, error) {
-	fm.muFiles.RLock()
-	defer fm.muFiles.RUnlock()
-
-	if page <= 0 || pageSize <= 0 {
-		slog.Warn("invalid page or pageSize for paginated file listing", "page", page, "pageSize", pageSize)
-		// Optionally return an error, but returning empty slice might be simpler for the caller
-		return []FileRecord{}, fmt.Errorf("invalid page number or page size")
-	}
-
-	records, exists := fm.fileRecords[areaID]
-	if !exists || len(records) == 0 {
-		// log.Printf("DEBUG: GetFilesForAreaPaginated called for non-existent, unloaded, or empty area ID %d.", areaID)
-		return []FileRecord{}, nil // Return empty slice if area empty or doesn't exist
-	}
-
-	totalFiles := len(records)
-	startIndex := (page - 1) * pageSize
-
-	// Check if start index is out of bounds
-	if startIndex >= totalFiles {
-		slog.Debug("requested page out of bounds for file listing", "page", page, "total", totalFiles, "pageSize", pageSize)
-		return []FileRecord{}, nil // Requested page is beyond the last file
-	}
-
-	endIndex := startIndex + pageSize
-	if endIndex > totalFiles {
-		endIndex = totalFiles // Adjust end index if it exceeds the total number of files
-	}
-
-	// Slice the records for the current page
-	pageRecords := records[startIndex:endIndex]
-
-	// Return a copy to prevent external modification
-	recordsCopy := make([]FileRecord, len(pageRecords))
-	copy(recordsCopy, pageRecords)
-
-	// log.Printf("DEBUG: GetFilesForAreaPaginated returning %d records for area %d, page %d", len(recordsCopy), areaID, page)
-	return recordsCopy, nil
-}
-
-// AddFileRecord adds a new file record to the specified area and saves.
-func (fm *FileManager) AddFileRecord(record FileRecord) error {
-	// Basic validation
-	if record.ID == uuid.Nil {
-		return fmt.Errorf("file record must have a valid ID")
-	}
-	if record.AreaID <= 0 {
-		return fmt.Errorf("file record must have a valid AreaID")
-	}
-	if record.Filename == "" {
-		return fmt.Errorf("file record must have a Filename")
-	}
-
-	fm.muAreas.RLock() // Check if area exists
-	_, areaExists := fm.fileAreas[record.AreaID]
-	fm.muAreas.RUnlock()
-	if !areaExists {
-		return fmt.Errorf("cannot add record to non-existent area ID %d", record.AreaID)
-	}
-
-	fm.muFiles.Lock()
-	defer fm.muFiles.Unlock()
-
-	// Check for duplicate filename within the same area?
-	// Or should uploads handle overwrites/renaming?
-	// For now, allow duplicates, but log a warning.
-	existingRecords := fm.fileRecords[record.AreaID]
-	for _, existing := range existingRecords {
-		if strings.EqualFold(existing.Filename, record.Filename) {
-			slog.Warn("adding file record with duplicate filename", "file", record.Filename, "area", record.AreaID)
-			break
-		}
-	}
-
-	fm.fileRecords[record.AreaID] = append(fm.fileRecords[record.AreaID], record)
-
-	// Save changes for this area - release lock during save
-	fm.muFiles.Unlock()
-	err := fm.saveFileRecords(record.AreaID)
-	fm.muFiles.Lock() // Re-acquire lock before returning
-
-	if err != nil {
-		// Attempt to rollback the in-memory addition? Complex.
-		slog.Error("failed to save file records after adding, in-memory state might be inconsistent", "file", record.Filename, "area", record.AreaID, "error", err)
-		// Maybe remove the last added record if save fails?
-		// fm.fileRecords[record.AreaID] = fm.fileRecords[record.AreaID][:len(fm.fileRecords[record.AreaID])-1]
-		return err // Propagate save error
-	}
-
-	slog.Info("added file record", "file", record.Filename, "id", record.ID, "area", record.AreaID)
-	return nil
-}
-
-// IncrementDownloadCount increments the download count for a file and saves.
-func (fm *FileManager) IncrementDownloadCount(fileID uuid.UUID) error {
-	fm.muFiles.Lock()
-	defer fm.muFiles.Unlock()
-
-	foundAreaID := -1
-	foundIndex := -1
-
-	// Find the file across all areas
-searchLoop:
-	for areaID, records := range fm.fileRecords {
-		for i := range records {
-			if records[i].ID == fileID {
-				foundAreaID = areaID
-				foundIndex = i
-				break searchLoop
-			}
-		}
-	}
-
-	if foundAreaID == -1 {
-		return fmt.Errorf("file record with ID %s not found", fileID)
-	}
-
-	// Increment count directly on the pointer within the slice
-	fm.fileRecords[foundAreaID][foundIndex].DownloadCount++
-	newCount := fm.fileRecords[foundAreaID][foundIndex].DownloadCount
-	filename := fm.fileRecords[foundAreaID][foundIndex].Filename
-
-	// Save changes for this area - release lock during save
-	fm.muFiles.Unlock()
-	err := fm.saveFileRecords(foundAreaID)
-	fm.muFiles.Lock() // Re-acquire lock before returning
-
-	if err != nil {
-		slog.Error("failed to save file records after incrementing download count, in-memory state might be inconsistent", "file", filename, "id", fileID, "error", err)
-		// Attempt rollback?
-		// fm.fileRecords[foundAreaID][foundIndex].DownloadCount--
-		return err
-	}
-
-	slog.Debug("incremented download count for file", "file", filename, "id", fileID, "count", newCount)
-	return nil
-}
-
-// UpdateFileRecord finds a file record by ID and applies the given update function.
-func (fm *FileManager) UpdateFileRecord(fileID uuid.UUID, updateFunc func(*FileRecord)) error {
-	fm.muFiles.Lock()
-	defer fm.muFiles.Unlock()
-
-	foundAreaID := -1
-	foundIndex := -1
-
-searchLoop:
-	for areaID, records := range fm.fileRecords {
-		for i := range records {
-			if records[i].ID == fileID {
-				foundAreaID = areaID
-				foundIndex = i
-				break searchLoop
-			}
-		}
-	}
-
-	if foundAreaID == -1 {
-		return fmt.Errorf("file record with ID %s not found", fileID)
-	}
-
-	updateFunc(&fm.fileRecords[foundAreaID][foundIndex])
-	filename := fm.fileRecords[foundAreaID][foundIndex].Filename
-
-	fm.muFiles.Unlock()
-	err := fm.saveFileRecords(foundAreaID)
-	fm.muFiles.Lock()
-
-	if err != nil {
-		slog.Error("failed to save file records after updating", "file", filename, "id", fileID, "error", err)
-		return err
-	}
-
-	slog.Debug("updated file record", "file", filename, "id", fileID)
-	return nil
-}
-
-// UpdateFileDescription updates the description of a file record.
-func (fm *FileManager) UpdateFileDescription(fileID uuid.UUID, description string) error {
-	return fm.UpdateFileRecord(fileID, func(r *FileRecord) {
-		r.Description = description
-	})
-}
-
-// DeleteFileRecord removes a file record by ID. If deleteFromDisk is true,
-// the physical file is also removed from the filesystem.
-func (fm *FileManager) DeleteFileRecord(fileID uuid.UUID, deleteFromDisk bool) error {
-	fm.muFiles.Lock()
-	defer fm.muFiles.Unlock()
-
-	foundAreaID := -1
-	foundIndex := -1
-	var foundFilename string
-
-searchLoop:
-	for areaID, records := range fm.fileRecords {
-		for i := range records {
-			if records[i].ID == fileID {
-				foundAreaID = areaID
-				foundIndex = i
-				foundFilename = records[i].Filename
-				break searchLoop
-			}
-		}
-	}
-
-	if foundAreaID == -1 {
-		return fmt.Errorf("file record with ID %s not found", fileID)
-	}
-
-	// If requested, delete from disk first — before touching metadata.
-	// This way a failure leaves metadata intact and the operation is retryable.
-	if deleteFromDisk {
-		fm.muAreas.RLock()
-		area, areaExists := fm.fileAreas[foundAreaID]
-		fm.muAreas.RUnlock()
-		if !areaExists {
-			return fmt.Errorf("internal inconsistency: area %d not found", foundAreaID)
-		}
-		absBasePath, err := filepath.Abs(fm.basePath)
-		if err != nil {
-			return fmt.Errorf("failed to get absolute base path: %w", err)
-		}
-		fullPath := filepath.Join(absBasePath, area.Path, filepath.Base(foundFilename))
-		if err := os.Remove(fullPath); err != nil && !os.IsNotExist(err) {
-			slog.Warn("failed to delete file from disk", "path", fullPath, "error", err)
-			return fmt.Errorf("failed to delete file from disk: %w", err)
-		}
-		slog.Info("deleted file from disk", "path", fullPath)
-	}
-
-	// Remove record from slice and persist.
-	records := fm.fileRecords[foundAreaID]
-	fm.fileRecords[foundAreaID] = append(records[:foundIndex], records[foundIndex+1:]...)
-
-	fm.muFiles.Unlock()
-	saveErr := fm.saveFileRecords(foundAreaID)
-	fm.muFiles.Lock()
-
-	if saveErr != nil {
-		slog.Error("failed to save file records after deleting", "file", foundFilename, "id", fileID, "error", saveErr)
-		return saveErr
-	}
-
-	slog.Info("deleted file record", "file", foundFilename, "id", fileID, "area", foundAreaID)
-	return nil
-}
-
-// MoveFileRecord moves a file record to a different area, renaming the file on disk.
-func (fm *FileManager) MoveFileRecord(fileID uuid.UUID, targetAreaID int) error {
-	fm.muAreas.RLock()
-	targetArea, targetExists := fm.fileAreas[targetAreaID]
-	fm.muAreas.RUnlock()
-	if !targetExists {
-		return fmt.Errorf("target area ID %d not found", targetAreaID)
-	}
-
-	fm.muFiles.Lock()
-	defer fm.muFiles.Unlock()
-
-	srcAreaID := -1
-	srcIndex := -1
-
-searchLoop:
-	for areaID, records := range fm.fileRecords {
-		for i := range records {
-			if records[i].ID == fileID {
-				srcAreaID = areaID
-				srcIndex = i
-				break searchLoop
-			}
-		}
-	}
-
-	if srcAreaID == -1 {
-		return fmt.Errorf("file record with ID %s not found", fileID)
-	}
-	if srcAreaID == targetAreaID {
-		return fmt.Errorf("file is already in area %d", targetAreaID)
-	}
-
-	record := fm.fileRecords[srcAreaID][srcIndex]
-
-	fm.muAreas.RLock()
-	srcArea, srcExists := fm.fileAreas[srcAreaID]
-	fm.muAreas.RUnlock()
-	if !srcExists {
-		return fmt.Errorf("internal inconsistency: source area %d not found", srcAreaID)
-	}
-
-	absBasePath, err := filepath.Abs(fm.basePath)
-	if err != nil {
-		return fmt.Errorf("failed to get absolute base path: %w", err)
-	}
-	safeFilename := filepath.Base(record.Filename)
-	srcPath := filepath.Join(absBasePath, srcArea.Path, safeFilename)
-	dstPath := filepath.Join(absBasePath, targetArea.Path, safeFilename)
-
-	// Guard against silently overwriting an existing file in the target area.
-	if _, err := os.Stat(dstPath); err == nil {
-		return fmt.Errorf("file %q already exists in target area %d", safeFilename, targetAreaID)
-	}
-
-	if err := os.Rename(srcPath, dstPath); err != nil {
-		return fmt.Errorf("failed to move file from %s to %s: %w", srcPath, dstPath, err)
-	}
-
-	// Update in-memory state.
-	srcRecords := fm.fileRecords[srcAreaID]
-	fm.fileRecords[srcAreaID] = append(srcRecords[:srcIndex], srcRecords[srcIndex+1:]...)
-	record.AreaID = targetAreaID
-	fm.fileRecords[targetAreaID] = append(fm.fileRecords[targetAreaID], record)
-
-	fm.muFiles.Unlock()
-	errSrc := fm.saveFileRecords(srcAreaID)
-	errDst := fm.saveFileRecords(targetAreaID)
-	fm.muFiles.Lock()
-
-	if errSrc != nil || errDst != nil {
-		// At least one metadata save failed. Roll back the filesystem rename so
-		// the file returns to the source directory and the operation is retryable.
-		if renameBackErr := os.Rename(dstPath, srcPath); renameBackErr != nil {
-			slog.Error("failed to roll back file rename after metadata save failure", "from", dstPath, "to", srcPath, "error", renameBackErr)
-		} else {
-			// Restore in-memory state.
-			tgtRecords := fm.fileRecords[targetAreaID]
-			fm.fileRecords[targetAreaID] = tgtRecords[:len(tgtRecords)-1]
-			record.AreaID = srcAreaID
-			fm.fileRecords[srcAreaID] = append(fm.fileRecords[srcAreaID], record)
-			// Re-persist both areas so disk reflects the restored in-memory state.
-			// A partial save (e.g. errSrc==nil but errDst!=nil) may have already
-			// written one side; re-saving corrects any such divergence.
-			fm.muFiles.Unlock()
-			if resaveErr := fm.saveFileRecords(srcAreaID); resaveErr != nil {
-				slog.Error("failed to re-save source area during rollback", "area", srcAreaID, "error", resaveErr)
-			}
-			if resaveErr := fm.saveFileRecords(targetAreaID); resaveErr != nil {
-				slog.Error("failed to re-save target area during rollback", "area", targetAreaID, "error", resaveErr)
-			}
-			fm.muFiles.Lock()
-		}
-		if errSrc != nil {
-			slog.Error("failed to save source area after moving", "area", srcAreaID, "file", safeFilename, "error", errSrc)
-			return errSrc
-		}
-		slog.Error("failed to save target area after moving", "area", targetAreaID, "file", safeFilename, "error", errDst)
-		return errDst
-	}
-
-	slog.Info("moved file", "file", safeFilename, "id", fileID, "from", srcAreaID, "to", targetAreaID)
-	return nil
-}
-
-// GetFilePath returns the full, absolute path to a file given its record ID.
-// It checks that the file exists and constructs the path safely.
-func (fm *FileManager) GetFilePath(fileID uuid.UUID) (string, error) {
-	fm.muFiles.RLock() // Need read lock to find the file record
-	defer fm.muFiles.RUnlock()
-	fm.muAreas.RLock() // Need read lock to get area path
-	defer fm.muAreas.RUnlock()
-
-	var foundArea *FileArea
-	var foundRecord *FileRecord
-
-searchLoop:
-	for areaID, records := range fm.fileRecords {
-		for i := range records {
-			if records[i].ID == fileID {
-				// Get corresponding area
-				area, areaExists := fm.fileAreas[areaID]
-				if !areaExists {
-					// Should not happen if data is consistent
-					return "", fmt.Errorf("internal inconsistency: area %d not found for file %s", areaID, fileID)
-				}
-				foundArea = area
-				foundRecord = &records[i] // Get pointer to the record
-				break searchLoop
-			}
-		}
-	}
-
-	if foundRecord == nil {
-		return "", fmt.Errorf("file record with ID %s not found", fileID)
-	}
-
-	// Construct path safely
-	// Base path should be absolute for security
-	absBasePath, err := filepath.Abs(fm.basePath)
-	if err != nil {
-		return "", fmt.Errorf("failed to get absolute base path: %w", err)
-	}
-	// Area path is relative to base path
-	// Filename should be just the base name
-	safeFilename := filepath.Base(foundRecord.Filename)
-	if safeFilename == "." || safeFilename == "/" || strings.Contains(safeFilename, "..") {
-		return "", fmt.Errorf("invalid filename in record: %s", foundRecord.Filename)
-	}
-
-	fullPath := filepath.Join(absBasePath, foundArea.Path, safeFilename)
-
-	// Final check: Ensure the resolved path is still within the intended base directory
-	if !strings.HasPrefix(fullPath, absBasePath) {
-		return "", fmt.Errorf("constructed file path '%s' is outside base directory '%s'", fullPath, absBasePath)
-	}
-
-	// Optional: Check if file actually exists on disk?
-	// if _, err := os.Stat(fullPath); os.IsNotExist(err) {
-	// 	return "", fmt.Errorf("file '%s' not found on disk at expected path '%s'", safeFilename, fullPath)
-	// }
-
-	return fullPath, nil
-}
-
-// GetAreaUploadPath returns the absolute filesystem path for an area's file directory.
-func (fm *FileManager) GetAreaUploadPath(areaID int) (string, error) {
-	fm.muAreas.RLock()
-	defer fm.muAreas.RUnlock()
-
-	area, exists := fm.fileAreas[areaID]
-	if !exists {
-		return "", fmt.Errorf("file area %d not found", areaID)
-	}
-
-	absBasePath, err := filepath.Abs(fm.basePath)
-	if err != nil {
-		return "", fmt.Errorf("failed to get absolute base path: %w", err)
-	}
-
-	fullPath := filepath.Join(absBasePath, area.Path)
-	if !strings.HasPrefix(fullPath, absBasePath) {
-		return "", fmt.Errorf("area path outside base directory")
-	}
-
-	return fullPath, nil
-}
-
-// IsSupportedArchive checks if the filename suggests a supported archive type.
-// Uses the central archivers.json configuration to determine supported formats.
-func (fm *FileManager) IsSupportedArchive(filename string) bool {
-	// Load the central archiver config. Defaults are used if archivers.json is missing.
-	arcCfg, err := archiver.LoadConfig("configs")
-	if err != nil {
-		slog.Warn("failed to load archivers config, falling back to .zip only", "error", err)
-		return strings.HasSuffix(strings.ToLower(filename), ".zip")
-	}
-	return arcCfg.IsSupported(filename)
 }
